fix norm and backspaces/delete
