+++ src/readline/cursor_view.c
@@ -6,7 +6,7 @@
 /*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/25 13:46:41 by lroussel          #+#    #+#             */
-/*   Updated: 2025/03/25 13:50:07 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/03/25 14:37:33 by lroussel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,4 +21,3 @@ void	show_cursor(void)
 {
 	write(0, SHOW_CURSOR, ft_strlen(SHOW_CURSOR));
 }
-
